Refactor Infinispan to delete cluster when startup failures
++ To move all the common functionality into a base class.
