+++ infinispan-client/src/test/java/io/quarkus/ts/infinispan/client/OperatorOpenShiftInfinispanCountersIT.java
@@ -5,27 +5,16 @@ import static org.hamcrest.CoreMatchers.is;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 import java.io.IOException;
-import java.nio.charset.Charset;
-import java.nio.charset.StandardCharsets;
-import java.nio.file.Files;
-import java.nio.file.Path;
-import java.nio.file.Paths;
 import java.util.concurrent.TimeUnit;
-
-import javax.inject.Inject;
 
 import org.apache.http.HttpStatus;
 import org.hamcrest.Matchers;
-import org.junit.jupiter.api.AfterAll;
-import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.MethodOrderer;
 import org.junit.jupiter.api.Order;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.TestMethodOrder;
-import org.junit.jupiter.api.condition.EnabledIfSystemProperty;
 
 import io.quarkus.test.bootstrap.RestService;
-import io.quarkus.test.bootstrap.inject.OpenShiftClient;
 import io.quarkus.test.scenarios.OpenShiftDeploymentStrategy;
 import io.quarkus.test.scenarios.OpenShiftScenario;
 import io.quarkus.test.services.QuarkusApplication;
@@ -33,34 +22,18 @@ import io.quarkus.test.utils.Command;
 
 @OpenShiftScenario(deployment = OpenShiftDeploymentStrategy.UsingOpenShiftExtension)
 @TestMethodOrder(MethodOrderer.OrderAnnotation.class)
-@EnabledIfSystemProperty(named = "ts.redhat.registry.enabled", matches = "true")
-public class OperatorOpenShiftInfinispanCountersIT {
-
-    private static final String ORIGIN_CLUSTER_NAME = "totally-random-infinispan-cluster-name";
-
-    private static final String TARGET_RESOURCES = "target/test-classes/";
-    private static final String CLUSTER_SECRET = "clientcert_secret.yaml";
-    private static final String CLUSTER_CONFIG = "infinispan_cluster_config.yaml";
-    private static final String CLUSTER_CONFIGMAP = "infinispan_cluster_configmap.yaml";
-    private static final String CONNECT_SECRET = "connect_secret.yaml";
-    private static final String TLS_SECRET = "tls_secret.yaml";
-
-    private static final String CLUSTER_NAMESPACE_NAME = "datagrid-cluster";
-    private static String NEW_CLUSTER_NAME = null;
-
-    @Inject
-    static OpenShiftClient ocClient;
+public class OperatorOpenShiftInfinispanCountersIT extends BaseOpenShiftInfinispanIT {
 
     @QuarkusApplication
     static RestService one = new RestService()
-            .onPreStart(OperatorOpenShiftInfinispanCountersIT::createInfinispanCluster);
+            .onPreStart(s -> createInfinispanCluster());
 
     @QuarkusApplication
     static RestService two = new RestService();
 
     /**
      * Simple check of connection to endpoints
-     *
+     * <p>
      * Expected values = 0
      */
     @Test
@@ -337,36 +310,6 @@ public class OperatorOpenShiftInfinispanCountersIT {
         assertEquals("Cache=2 Client=2", secondClientCounters);
     }
 
-    private static void createInfinispanCluster(io.quarkus.test.bootstrap.Service service) {
-        applyYaml(CLUSTER_SECRET);
-        applyYaml(CONNECT_SECRET);
-        applyYaml(TLS_SECRET);
-
-        // there should be unique name for every created infinispan cluster to be able parallel runs
-        NEW_CLUSTER_NAME = ocClient.project() + "-infinispan-cluster";
-
-        // rename infinispan cluster and configmap
-        adjustYml(CLUSTER_CONFIG, ORIGIN_CLUSTER_NAME, NEW_CLUSTER_NAME);
-        applyYaml(CLUSTER_CONFIG);
-        adjustYml(CLUSTER_CONFIGMAP, ORIGIN_CLUSTER_NAME, NEW_CLUSTER_NAME);
-        applyYaml(CLUSTER_CONFIGMAP);
-
-        try {
-            new Command("oc", "-n", CLUSTER_NAMESPACE_NAME, "wait", "--for", "condition=wellFormed", "--timeout=300s",
-                    "infinispan/" + NEW_CLUSTER_NAME).runAndWait();
-        } catch (Exception e) {
-            Assertions.fail("Fail to wait Infinispan resources to start. Caused by: " + e.getMessage());
-        }
-    }
-
-    @AfterAll
-    public static void deleteInfinispanCluster() {
-        deleteYaml(CLUSTER_CONFIGMAP);
-        deleteYaml(CLUSTER_CONFIG);
-        adjustYml(CLUSTER_CONFIG, NEW_CLUSTER_NAME, ORIGIN_CLUSTER_NAME);
-        adjustYml(CLUSTER_CONFIGMAP, NEW_CLUSTER_NAME, ORIGIN_CLUSTER_NAME);
-    }
-
     /**
      * Setting the cache counter value to 0 from provided client url address.
      * At the end, the cache value is tested that it is actually 0.
@@ -469,44 +412,5 @@ public class OperatorOpenShiftInfinispanCountersIT {
         applyYaml(CLUSTER_CONFIG);
         new Command("oc", "-n", CLUSTER_NAMESPACE_NAME, "wait", "--for", "condition=wellFormed", "--timeout=360s",
                 "infinispan/" + NEW_CLUSTER_NAME).runAndWait();
-    }
-
-    /**
-     * Replacing values in the provided YAML file and then apply it.
-     */
-    private static void adjustYml(String yamlFile, String originString, String newString) {
-        try {
-            Path yamlPath = Paths.get(TARGET_RESOURCES + yamlFile);
-            Charset charset = StandardCharsets.UTF_8;
-
-            String yamlContent = new String(Files.readAllBytes(yamlPath), charset);
-            yamlContent = yamlContent.replace(originString, newString);
-            Files.write(yamlPath, yamlContent.getBytes(charset));
-        } catch (IOException ex) {
-            Assertions.fail("Fail to adjust YAML file. Caused by: " + ex.getMessage());
-        }
-    }
-
-    /**
-     * Apply the YAML file.
-     */
-    private static void applyYaml(String yamlFile) {
-        try {
-            new Command("oc", "apply", "-f", Paths.get(TARGET_RESOURCES + yamlFile).toString()).runAndWait();
-        } catch (Exception e) {
-            Assertions.fail("Failed to apply YAML file. Caused by: " + e.getMessage());
-        }
-    }
-
-    /**
-     *
-     * Delete the YAML file.
-     */
-    private static void deleteYaml(String yamlFile) {
-        try {
-            new Command("oc", "delete", "-f", Paths.get(TARGET_RESOURCES + yamlFile).toString()).runAndWait();
-        } catch (Exception e) {
-            Assertions.fail("Failed to delete YAML file. Caused by: " + e.getMessage());
-        }
     }
 }
